requirements.txt
streamlit>=1.36
numpy>=1.24
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
scipy>=1.11
fastdtw>=0.3.4
pillow>=10.3
matplotlib>=3.8
